+++ test/sys/test_sockopt.rs
@@ -1270,3 +1270,34 @@ pub fn test_so_attach_reuseport_cbpf() {
         assert_eq!(e, nix::errno::Errno::ENOPROTOOPT);
     });
 }
+
+#[test]
+#[cfg(not(any(
+    target_os = "fuchsia",
+    target_os = "haiku",
+    target_os = "netbsd",
+    target_os = "openbsd"
+)))]
+fn test_source_membership_opts() {
+    let fd = socket(
+        AddressFamily::Inet,
+        SockType::Datagram,
+        SockFlag::empty(),
+        SockProtocol::Udp,
+    )
+    .unwrap();
+    let request = nix::sys::socket::IpSourceMembershipRequest::new(
+        "224.0.0.0".parse().unwrap(),
+        "127.0.0.1".parse().unwrap(),
+        None,
+    );
+
+    setsockopt(fd, sockopt::IpDropSourceMembership, &request)
+        .expect_err("drop not exist multicast membership should fail");
+
+    setsockopt(fd, sockopt::IpAddSourceMembership, &request)
+        .expect("add source specific multicast membership should succeed");
+
+    setsockopt(fd, sockopt::IpDropSourceMembership, &request)
+        .expect("drop source specific multicast membership should succeed");
+}